requirements.txt
blinker==1.6.3
click==8.1.7
colorama==0.4.6
DateTime==5.2
Flask==3.0.0
greenlet==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
SQLAlchemy==2.0.22
typing_extensions==4.8.0
Werkzeug==3.0.0
zope.interface==6.1